pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixel-police"
version = "0.1.0"
description = "Visual regression testing: screenshot comparison with before/after diff"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "pixel-police contributors"}]
keywords = ["visual-regression", "testing", "screenshot", "diff", "playwright"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "click>=8.0",
    "Pillow>=10.0",
    "rich>=13.0",
    "playwright>=1.40",
]

[project.scripts]
pixel-police = "pixel_police.cli:cli"

[project.urls]
Homepage = "https://github.com/pixel-police/pixel-police"
Repository = "https://github.com/pixel-police/pixel-police"

[tool.setuptools.packages.find]
include = ["pixel_police*"]
